Handled python version support

+++ modules/waitforupdates.py
@@ -1,3 +1,20 @@
+#
+# Copyright 2016 BMC Software, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# 
+
+
 # Monitoring property changes to objects of one or more types
 from modules import serviceutil
 from pyVim.connect import SmartConnect, Disconnect
@@ -103,7 +120,7 @@ def monitor_property_changes(si, propspec, self, discoverySelfInstance, iteratio
 
     pc = si.content.propertyCollector
     make_property_collector(pc, si.content.rootFolder, propspec, self)
-    waitforUpdateTimeInseconds = self.params['discoveryInterval'] / 1000
+    waitforUpdateTimeInseconds = int(self.params['discoveryInterval'] / 1000)
     waitopts = make_wait_options(waitforUpdateTimeInseconds)
 
     version = ''
